feat: BigDecimal을 Double로 수정

+++ src/main/java/com/poortorich/chart/response/CategoryChart.java
@@ -1,16 +1,15 @@
 package com.poortorich.chart.response;
 
-import java.math.BigDecimal;
 import lombok.Builder;
 import lombok.Data;
 
 @Data
 @Builder
 public class CategoryChart {
-    
+
     private Long id;
     private String color;
     private String name;
-    private BigDecimal rate;
+    private Double rate;
     private Long amount;
 }
